requirements.txt
openai
faiss-cpu
python-dotenv
mcp
langchain-text-splitters
sentence-transformers
